[libc++] Make sure we use the libdispatch backend on Apple platforms
Summary:
The Apple.cmake cache wasn't set up properly, so we wouldn't enable
the libdispatch backend by default on Apple platforms. This patch
fixes the issue and adds a test.

We drop the usage of std::vector in libdispatch.h to avoid changing
the transitive includes only on Apple platforms.

Reviewers: philnik

Subscribers: