[clang] Move the Clang CI jobs off of the libc++ builders
Summary:
The libc++ builders do not have the capacity to serve both libc++ and
Clang CI needs. We initially used the libc++ builders when adding
pre-commit CI for Clang as an experiment, however this setup has been
problematic for libc++ for a long time. It makes libc++ CI turnaround
way too slow, which has been a frequent source of frustration.

We didn't want to make this change before the release to avoid rocking
the boat too much in such an important time, however now that the
release has been cut, we need to make this change. This will shift
the traffic for Clang pre-commit CI from the libc++ builders to the
builders that have been used for Clang pre-commit CI defined outside
of Clang's BuildKite pipeline.

Reviewers:

Subscribers: aaron.ballman