[clang] Fix the pre-commit CI pipeline after changes to libc++'s own CI pipeline
We made some changes to the libc++ CI pipeline that take for granted
that we're running on libc++'s own Docker images. This was necessary for
a temporary period until widely-used tools update to a version that can
handle C++20 modules.

However, this had the unintended consequence of breaking the Clang CI
pipeline, which used the libc++ CI scripts as an implementation detail.
Instead, decouple the Clang CI pipeline from the libc++ build scripts.