[runtimes] Add the ability to include out of tree runtimes
Summary:
LLVM_EXTERNAL_RUNTIMES is similar to LLVM_EXTERNAL_PROJECTS: the
source directory for the external runtime is resolved through the
LLVM_EXTERNAL_<proj>_SOURCE_DIR variable.

This makes it easier to experiment with additional runtimes locally
and provides a relatively low maintenance hook for downstreams to
add their own runtimes outside of the LLVM tree.

Reviewers: phosek

Subscribers: